Change .call QMDensityEig Layer
I propose changing the .call method in QMDensityEig layer, the reason is that originally the layer computes tr ( |psi><psi|(rho_train)|psi><psi| ) which is equivalent to <psi|(rho_train)|psi> , this reduces the complexity of the algorithm. I have checked that the outputs of the .call layer is the same between the original method and the proposed method for random features and complex random features.